Extract area template functions into an areas Jinja2 extension
