+++ homeassistant/helpers/template/extensions/__init__.py
@@ -1,5 +1,6 @@
 """Home Assistant template extensions."""
 
+from .areas import AreaExtension
 from .base64 import Base64Extension
 from .collection import CollectionExtension
 from .crypto import CryptoExtension
@@ -11,6 +12,7 @@ from .regex import RegexExtension
 from .string import StringExtension
 
 __all__ = [
+    "AreaExtension",
     "Base64Extension",
     "CollectionExtension",
     "CryptoExtension",
